+++ front-end/Android/app/src/main/AndroidManifest.xml
@@ -13,7 +13,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".arch.view.CustomDetailsActivity"></activity>
+        <activity android:name=".view.community.CommunityWriteActivity"></activity>
+        <activity android:name=".arch.view.CustomDetailsActivity" />
         <activity android:name=".arch.view.ProfileActivity" />
         <activity android:name=".arch.view.FavoriteInitTagActivity" />
         <activity android:name=".view.account.FindPasswordActivity" />
