+++ front-end/Android/app/src/main/java/com/dnd/kindit/view/community/CommunityWriteActivity.kt
@@ -1,0 +1,12 @@
+package com.dnd.kindit.view.community
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.dnd.kindit.R
+
+class CommunityWriteActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_community_write)
+    }
+}
