+++ front-end/Android/app/src/main/res/layout/card_custom_detail_tag.xml
@@ -9,7 +9,7 @@
     <LinearLayout
         android:layout_width="80dp"
         android:layout_height="25dp"
-        android:background="@drawable/sendmail"
+        android:background="@drawable/write_button_upload"
         android:gravity="center"
         android:orientation="vertical"
         app:layout_constraintBottom_toBottomOf="parent"
